// src/constants.d.ts
export declare const AUTO_DECREASE_AMOUNT = 2;
export declare const AUTO_DECREASE_RATE = 100;
export declare const AUTO_INCREASE_AMOUNT = 10;
export declare const MAXIMUM_THERMOMETER_HEIGHT = 100;
export declare const EXPECTED_MAXIMUM_PERCENTAGE = 100;
export declare const KEYS_TO_HOLD: string[];
export declare const KEY_TO_PRESS = "r";
// ...
export declare const TRIAL_DURATION = 5000;
export declare const PARAMETER_COMBINATIONS: {
    reward: number[];
    bounds: number[];
}[];
export declare const PARAMETER_COMBINATIONS_TOTAL: number;
export declare const GO_DURATION = 500;
export declare const SUCCESS_SCREEN_DURATION = 500;
export declare const COUNTDOWN_TIME = 3;
export declare const PREMATURE_KEY_RELEASE_ERROR_TIME = 1000;
export declare const KEY_TAPPED_EARLY_ERROR_TIME = 3000;
// ...
export declare const PREMATURE_KEY_RELEASE_ERROR_MESSAGE: string;
export declare const LIKERT_PREAMBLE: string;
export declare const FAILED_MINIMUM_DEMO_TAPS_MESSAGE: string;
export declare const HOLD_KEYS_MESSAGE: string;
export declare const KEY_TAPPED_EARLY_MESSAGE: string;